+++ test/e2e/busted__vra_core.lua
@@ -142,7 +142,7 @@ describe("busted", function()
     end)
 
     describe("Check VRA Announce:", function()
-        before_each(function()
+		before_each(function()
             -- Reset before each test
             fakeCombatLogEvent = {}
         end)
@@ -151,25 +151,32 @@ describe("busted", function()
 			assert.spy(f)[assertion](...)
 		end
 
-		local function checkSpellCastEvent(cleu, isHarmful, onlySelf, expected_result_string, assert_func)
-			it(expected_result_string, function()
+		local function checkSpellCastEvent(cleu, isHarmfulSpell, onlySelf, expected_result_string, assert_func)
+			isHarmful = isHarmfulSpell
+			describe("", function()
+				it(expected_result_string, function()
 				setMockCombatLogEntry(cleu)
 				local play = spy.on(addon, "playSpell")
 
 				addon.profile.general.area["STUB"].onlySelf = onlySelf
 				addon:COMBAT_LOG_EVENT_UNFILTERED("COMBAT_LOG_EVENT_UNFILTERED")
 				spy_assert(play, assert_func)
+				play:revert()
+				end)
 			end)
 		end
 
-		local function checkAuraAppEvent(cleu, isHarmful, onlySelf, expected_result_string, assert_func)
-			it(expected_result_string, function()
+		local function checkAuraAppEvent(cleu, onlySelf, expected_result_string, assert_func)
+			describe("", function()
+				it(expected_result_string, function()
 				setMockCombatLogEntry(cleu)
 				local play = spy.on(addon, "playSpell")
 
 				addon.profile.general.area["STUB"].onlySelf = onlySelf
 				addon:COMBAT_LOG_EVENT_UNFILTERED("COMBAT_LOG_EVENT_UNFILTERED")
 				spy_assert(play, assert_func)
+				play:revert()
+				end)
 			end)
 		end
 
@@ -179,58 +186,19 @@ describe("busted", function()
 			-- SPELL_AURA_APPLIED should only play if "onlySelf" taken
             local logStringEvokerOnMeCast = "123456.78, SPELL_CAST_SUCCESS, false, Player-1234-456, CasterName, 1298, 0, Player-1234-123, TargetName, 1297, 0, 406732, Spatial Paradox, 64"
 			local logStringEvokerOnMeAura = "123456.78, SPELL_AURA_APPLIED, false, Player-1234-456, CasterName, 1298, 0, Player-1234-123, TargetName, 1297, 0, 406732, Spatial Paradox, 64"
-            isHarmful = false
 
 			describe("the spell cast should:", function()
 				checkSpellCastEvent(logStringEvokerOnMeCast, false, true, "not play if only self = true", "not_called")
 				checkSpellCastEvent(logStringEvokerOnMeCast, false, false, "not play if only self = false", "not_called")
-				--[[
-				it("not play if only self = true", function()
-					setMockCombatLogEntry(logStringEvokerOnMeCast)
-					local play = spy.on(addon, "playSpell")
-
-					addon.profile.general.area["STUB"].onlySelf = true
-					addon:COMBAT_LOG_EVENT_UNFILTERED("COMBAT_LOG_EVENT_UNFILTERED")
-					assert.spy(play).was.not_called()
-				end)
-				it("not play if only self = false", function()
-					setMockCombatLogEntry(logStringEvokerOnMeCast)
-					local play = spy.on(addon, "playSpell")
-					addon.profile.general.area["STUB"].onlySelf = false
-					addon:COMBAT_LOG_EVENT_UNFILTERED("COMBAT_LOG_EVENT_UNFILTERED")
-					assert.spy(play).was.not_called()
-				end)
-				]]--
 			end)
 
 			describe("the aura application should:", function()
-				checkAuraAppEvent(logStringEvokerOnMeAura, false, true, "play if only self = true", "called")
-				checkAuraAppEvent(logStringEvokerOnMeAura, false, false, "play, if only self = false", "called")
-				--[[
-				it("play if only self = true", function()
-					setMockCombatLogEntry(logStringEvokerOnMeAura)
-                local play = spy.on(addon, "playSpell")
-
-                addon.profile.general.area["STUB"].onlySelf = true
-                addon:COMBAT_LOG_EVENT_UNFILTERED("COMBAT_LOG_EVENT_UNFILTERED")
-                assert.spy(play).was.called()
-				end)
-
-				it("play, if only self = false", function()
-					setMockCombatLogEntry(logStringEvokerOnMeAura)
-                local play = spy.on(addon, "playSpell")
-
-                addon.profile.general.area["STUB"].onlySelf = false
-                addon:COMBAT_LOG_EVENT_UNFILTERED("COMBAT_LOG_EVENT_UNFILTERED")
-                assert.spy(play).was.called()
-				end)
-				]]--
+				checkAuraAppEvent(logStringEvokerOnMeAura, false, "play if only self = true", "called")
+				checkAuraAppEvent(logStringEvokerOnMeAura, false, "play, if only self = false", "called")
 			end)
         end)
 
-		--[[
-
-        describe("Partymember casts a spell (with aura) on: -> Partymember / Not the Player <-:", function()
+        describe("Partymember casts a friendly spell (with aura) on: -> Partymember / Not the Player <-:", function()
             -- Evoker in our group casts Spatial Paradox on "someone"
 			-- SPELL_CAST_SUCCESS should play if target = not player
 			-- SPELL_AURA_APPLIED should never play
@@ -239,45 +207,18 @@ describe("busted", function()
 			local logStringEvokerOnThirdAura = "123456.78, SPELL_AURA_APPLIED, false, Player-1234-456, CasterName, 1298, 0, Player-1234-444, TargetName, 1298, 0, 406732, Spatial Paradox, 64"
 
 			describe("the spell cast should:", function()
-				it("not play if only self = true", function()
-					setMockCombatLogEntry(logStringEvokerOnThirdCast)
-					local play = spy.on(addon, "playSpell")
-
-					addon.profile.general.area["STUB"].onlySelf = true
-					addon:COMBAT_LOG_EVENT_UNFILTERED("COMBAT_LOG_EVENT_UNFILTERED")
-					assert.spy(play).was.not_called()
-				end)
-				it("not play if only self = false", function()
-					setMockCombatLogEntry(logStringEvokerOnThirdCast)
-					local play = spy.on(addon, "playSpell")
-					addon.profile.general.area["STUB"].onlySelf = false
-					addon:COMBAT_LOG_EVENT_UNFILTERED("COMBAT_LOG_EVENT_UNFILTERED")
-					assert.spy(play).was.not_called()
-				end)
+				checkSpellCastEvent(logStringEvokerOnThirdCast, false, true, "not play if only self = true", "not_called")
+				checkSpellCastEvent(logStringEvokerOnThirdCast, false, false, "not play if only self = false", "not_called")
 			end)
 
 			describe("the aura application should:", function()
-				it("not play if only self = true", function()
-					setMockCombatLogEntry(logStringEvokerOnThirdAura)
-                local play = spy.on(addon, "playSpell")
-
-                addon.profile.general.area["STUB"].onlySelf = true
-                addon:COMBAT_LOG_EVENT_UNFILTERED("COMBAT_LOG_EVENT_UNFILTERED")
-                assert.spy(play).was.not_called()
-				end)
-
-				it("play, if only self = false", function()
-					setMockCombatLogEntry(logStringEvokerOnThirdAura)
-                local play = spy.on(addon, "playSpell")
-
-                addon.profile.general.area["STUB"].onlySelf = false
-                addon:COMBAT_LOG_EVENT_UNFILTERED("COMBAT_LOG_EVENT_UNFILTERED")
-                assert.spy(play).was.called()
-				end)
+				checkAuraAppEvent(logStringEvokerOnThirdAura, true, "not play if only self = true", "not_called")
+				checkAuraAppEvent(logStringEvokerOnThirdAura, false, "play, if only self = false", "called")
 			end)
         end)
 
-		describe("Partymember casts a friendly spell (with aura) on: -> NPC Target <-:", function()
+
+		describe("Partymember casts a spell (with aura) on: -> NPC Target <-:", function()
 		-- Evoker in our group casts Spatial Paradox, but has the boss / create in target
 		-- Fallbacks:
 		-- --> 1. We get the Aura
@@ -288,101 +229,48 @@ describe("busted", function()
 		local logStringEvokerOnEnemyAura_Other = "123456.78, SPELL_AURA_APPLIED, false, Player-1234-444, CasterName, 1298, 0, Player-1234-555, TargetName, 1298, 0, 406732, Spatial Paradox, 64"
 
 			describe("the spell cast should:", function()
-				it ("always play, if it is a harmfull spell", function()
-					isHarmful = true
-					setMockCombatLogEntry(logStringEvokerOneEnemyCast)
-					local play = spy.on(addon, "playSpell")
-					addon.profile.general.area["STUB"].onlySelf = true
-					addon:COMBAT_LOG_EVENT_UNFILTERED("COMBAT_LOG_EVENT_UNFILTERED")
-					assert.spy(play).was.called()
-
-					addon.profile.general.area["STUB"].onlySelf = true
-					addon:COMBAT_LOG_EVENT_UNFILTERED("COMBAT_LOG_EVENT_UNFILTERED")
-					assert.spy(play).was.called(2)
+				describe("always play if it is a harmful spell:", function()
+					checkSpellCastEvent(logStringEvokerOneEnemyCast, true, false, "only self = false", "called")
+					checkSpellCastEvent(logStringEvokerOneEnemyCast, true, true, "only self = true", "called")
 				end)
-				it("not play if only self = true", function()
-					isHarmful = false
-					setMockCombatLogEntry(logStringEvokerOneEnemyCast)
-					local play = spy.on(addon, "playSpell")
-
-					addon.profile.general.area["STUB"].onlySelf = true
-					addon:COMBAT_LOG_EVENT_UNFILTERED("COMBAT_LOG_EVENT_UNFILTERED")
-					assert.spy(play).was.not_called()
-				end)
-				it("not play if only self = false", function()
-					isHarmful = false
-					setMockCombatLogEntry(logStringEvokerOneEnemyCast)
-					local play = spy.on(addon, "playSpell")
-					addon.profile.general.area["STUB"].onlySelf = false
-					addon:COMBAT_LOG_EVENT_UNFILTERED("COMBAT_LOG_EVENT_UNFILTERED")
-					assert.spy(play).was.not_called()
+				describe("never play, if it is helpful spell:", function()
+					checkSpellCastEvent(logStringEvokerOneEnemyCast, false, true, "only self = true", "not_called")
+					checkSpellCastEvent(logStringEvokerOneEnemyCast, false, false, "only self = false", "not_called")
 				end)
 			end)
 
 			describe("the aura application:", function()
 				-- Fallback #1 -> We get the Aura
 				describe("is applied to us and it should:", function()
-					it("play if only self = true", function()
-						setMockCombatLogEntry(logStringEvokerOnEnemyAura_Me)
-						local play = spy.on(addon, "playSpell")
-						addon.profile.general.area["STUB"].onlySelf = true
-						addon:COMBAT_LOG_EVENT_UNFILTERED("COMBAT_LOG_EVENT_UNFILTERED")
-						assert.spy(play).was.called()
-					end)
-
-					it("play, if only self = false", function()
-						setMockCombatLogEntry(logStringEvokerOnEnemyAura_Me)
-						local play = spy.on(addon, "playSpell")
-						addon.profile.general.area["STUB"].onlySelf = false
-						addon:COMBAT_LOG_EVENT_UNFILTERED("COMBAT_LOG_EVENT_UNFILTERED")
-						assert.spy(play).was.called()
-					end)
+					checkAuraAppEvent(logStringEvokerOnEnemyAura_Me, true, "play if only self = true", "called")
+					checkAuraAppEvent(logStringEvokerOnEnemyAura_Me, false, "play, if only self = false", "called")
 				end)
 				-- Fallback #2 -> Another Heal in Party gets the Aura
 				describe("is applied to party member and it should:", function()
-					it("not play if only self = true", function()
-						setMockCombatLogEntry(logStringEvokerOnEnemyAura_Other)
-						local play = spy.on(addon, "playSpell")
-						addon.profile.general.area["STUB"].onlySelf = true
-						addon:COMBAT_LOG_EVENT_UNFILTERED("COMBAT_LOG_EVENT_UNFILTERED")
-						assert.spy(play).was.not_called()
-					end)
-
-					it("play, if only self = false", function()
-						setMockCombatLogEntry(logStringEvokerOnEnemyAura_Other)
-						local play = spy.on(addon, "playSpell")
-						addon.profile.general.area["STUB"].onlySelf = false
-						addon:COMBAT_LOG_EVENT_UNFILTERED("COMBAT_LOG_EVENT_UNFILTERED")
-						assert.spy(play).was.called()
-					end)
+					checkAuraAppEvent(logStringEvokerOnEnemyAura_Other, true, "not play, if only self = true", "not_called")
+					checkAuraAppEvent(logStringEvokerOnEnemyAura_Other, false, "play, if only self = false", "called")
 				end)
 
 			end)
 		end)
+
 
 		describe("The player casts a harmful spell on: -> NPC Target <-:", function()
 			local logStringPlayerOnThird = "123456.78, SPELL_CAST_SUCCESS, false, Player-1234-123, CasterName, 1297, 0, Creature-1234-444, TargetName, 68168, 0, 322109, Touch of Death, 1"
-			isHarmful = true
+
 			describe("the spellcast should:", function()
-				it("not play a sound, if -> Watch for Own Abilities <- NOT taken", function()
-					setMockCombatLogEntry(logStringPlayerOnThird)
-                	local play = spy.on(addon, "playSpell")
-
-                	addon.profile.general.watchFor = 6 -- COMBATLOG_OBJECT_AFFILIATION_PARTY + COMBATLOG_OBJECT_AFFILIATION_RAID
-                	addon:COMBAT_LOG_EVENT_UNFILTERED("COMBAT_LOG_EVENT_UNFILTERED")
-                	assert.spy(play).was.not_called()
+				describe("not play: ", function()
+					-- COMBATLOG_OBJECT_AFFILIATION_PARTY + COMBATLOG_OBJECT_AFFILIATION_RAID
+					addon.profile.general.watchFor = 6
+					checkSpellCastEvent(logStringPlayerOnThird, true, false, "-> Watch for Own Abilities <- NOT taken", "not_called")
 				end)
-				it("play, if -> Watch for Own Abilities <- taken", function()
-					setMockCombatLogEntry(logStringPlayerOnThird)
-                	local play = spy.on(addon, "playSpell")
 
-					addon.profile.general.watchFor = 7 -- COMBATLOG_OBJECT_AFFILIATION_MINE + COMBATLOG_OBJECT_AFFILIATION_PARTY + COMBATLOG_OBJECT_AFFILIATION_RAID
-					addon:COMBAT_LOG_EVENT_UNFILTERED("COMBAT_LOG_EVENT_UNFILTERED")
-					assert.spy(play).was.called()
+				describe("play:", function()
+					-- COMBATLOG_OBJECT_AFFILIATION_MINE + COMBATLOG_OBJECT_AFFILIATION_PARTY + COMBATLOG_OBJECT_AFFILIATION_RAID
+					addon.profile.general.watchFor = 7
+					checkSpellCastEvent(logStringPlayerOnThird, true, false, "-> Watch for Own Abilities <- taken", "called")
 				end)
 			end)
-
 		end)
-	]]--
     end)
 end)