+++ changelog.md
@@ -1,7 +1,8 @@
 # Vocal Raid Assistant
 
 ## 3.9.0
-- Add fix for wrong announce when casting externals on NPCs. (e.g. Spatial Paradox "no healer" fallback)
+- Detect Externals via SPELL_AURA_APPLIED instead of SPELL_CAST_SUCCESS + Target Check.
+- This fixes the wrong announce when casting externals on NPCs. (e.g. Spatial Paradox "no healer" fallback)
 
 ## 3.8.1
 - Bump toc for 11.1.5